feat(loading): replace current loader with new svgs
BREAKING CHANGE: as above

DOOR-165
