+++ src/components/Loading/Loading.stories.tsx
@@ -1,11 +1,6 @@
 import React from 'react';
-import { States } from 'storybook-states';
-import { Loading, LoadingProps } from './Loading';
+import { Loading } from './Loading';
 
 export default { title: 'Components|Loading' };
 
-export const Default = () => (
-  <States<LoadingProps> states={[{ speed: 100 }, { speed: 250 }, { speed: 500 }]}>
-    <Loading />
-  </States>
-);
+export const Default = () => <Loading />;
