+++ src/components/Loading/Loading.tsx
@@ -1,29 +1,25 @@
 import React, { useEffect } from 'react';
 import { useCursor } from 'use-cursor';
 import classNames from 'classnames';
-import { colors } from '@moda/tokens';
-import { Brancusi, Bradley, Oval, Keyhole, Mirror } from '../Shape';
+import { LoadingTriangle, LoadingDiamond, LoadingWindow, LoadingCircle } from '../Shape';
 import './Loading.scss';
 
-export type LoadingProps = React.HTMLAttributes<HTMLDivElement> & {
-  speed?: number;
-};
+export type LoadingProps = React.HTMLAttributes<HTMLDivElement>;
 
 const FRAMES = [
-  <Brancusi key='brancusi' style={{ fill: colors.all.brick }} />,
-  <Bradley key='bradley' style={{ fill: colors.all.fuchsia }} />,
-  <Oval key='oval' style={{ fill: colors.all['cornflower-blue'] }} />,
-  <Keyhole key='keyhole' style={{ fill: colors.all.seafoam }} />,
-  <Mirror key='mirror' style={{ fill: colors.all.goldenrod }} />,
+  <LoadingTriangle key='triangle' />,
+  <LoadingDiamond key='diamond' />,
+  <LoadingWindow key='window' />,
+  <LoadingCircle key='circle' />
 ];
 
-export const Loading: React.FC<LoadingProps> = ({ className, speed = 250, ...rest }) => {
+export const Loading: React.FC<LoadingProps> = ({ className, ...rest }) => {
   const { index, handleNext } = useCursor({ max: FRAMES.length });
 
   useEffect(() => {
-    const interval = setInterval(handleNext, speed);
+    const interval = setInterval(handleNext, 1000);
     return () => clearInterval(interval);
-  }, [handleNext, speed]);
+  }, [handleNext]);
 
   return (
     <div className={classNames('Loading', className)} {...rest}>
