--- src/components/MiniLoading/MiniLoading.scss
@@ -1,30 +1,0 @@
-@import '~om';
-
-.MiniLoading {
-  padding: space(5);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-  width: 100%;
-  transition: opacity 1000ms;
-  .LoadingTriangle,
-  .LoadingDiamond,
-  .LoadingWindow,
-  .LoadingCircle {
-    width: space(7);
-    height: space(7);
-    opacity: 1;
-    animation: fade 1s linear;
-  }
-
-  @keyframes fade {
-    0%,
-    100% {
-      opacity: 0;
-    }
-    50% {
-      opacity: 1;
-    }
-  }
-}
