--- src/components/MiniLoading/MiniLoading.stories.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { MiniLoading } from './MiniLoading';
-
-export default { title: 'Components|MiniLoading' };
-
-export const Default = () => <MiniLoading />;
