--- src/components/MiniLoading/index.ts
@@ -1,1 +1,0 @@
-export * from './MiniLoading';
